requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
boto3==1.36.23
botocore==1.36.23
brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dacite==1.7.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
docker-pycreds==0.4.0
entrypoints==0.4
executing==2.2.0
fastjsonschema==2.21.1
fiftyone==1.3.0
fiftyone-brain==0.19.0
fiftyone-db==1.1.7
filelock==3.17.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
gitpython==3.1.44
glob2==0.7
graphql-core==3.2.6
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.1
hydra-core==1.3.2
hypercorn==0.17.3
hyperframe==6.1.0
idna==2.10
imageio==2.37.0
inflate64==1.0.1
ipykernel==6.29.5
ipython==8.32.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==3.5.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.16.3
jupyterlab-widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
lazy-loader==0.4
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multidict==6.1.0
multivolumefile==0.2.3
narwhals==1.27.1
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==6.5.7
notebook-shim==0.2.4
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py7zr==0.22.0
pybcj==1.0.3
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pydash==8.0.5
pygments==2.19.1
pymongo==4.9.2
pynndescent==0.5.13
pyparsing==3.2.1
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytorch-lightning==2.5.0.post0
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.2.1
pyzstd==0.16.2
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
requests==2.25.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rtree==1.3.0
s3transfer==0.11.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
send2trash==1.8.3
sentry-sdk==2.22.0
setproctitle==1.3.4
setuptools==75.8.0
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.45.3
strawberry-graphql==0.260.2
sympy==1.13.1
tabulate==0.9.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
tomli==2.2.1
torch==2.6.0+cu118
torchattacks==3.5.1
torchmetrics==1.6.1
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==1.26.20
voxel51-eta==0.14.0
wandb==0.19.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.3
